Extract callback configuration to a dedicated trait
