+++ src/Consumers/ConsumerBuilder.php
@@ -5,6 +5,7 @@ namespace Junges\Kafka\Consumers;
 use Closure;
 use InvalidArgumentException;
 use Junges\Kafka\Commit\Contracts\CommitterFactory;
+use Junges\Kafka\Concerns\InteractsWithConfigCallbacks;
 use Junges\Kafka\Config\BatchConfig;
 use Junges\Kafka\Config\Config;
 use Junges\Kafka\Config\NullBatchConfig;
@@ -18,6 +19,8 @@ use Junges\Kafka\Support\Timer;
 
 class ConsumerBuilder implements ConsumerBuilderContract
 {
+    use InteractsWithConfigCallbacks;
+
     protected array $topics;
     protected int $commit;
     protected ?string $groupId;
