+++ src/Contracts/ConsumerBuilder.php
@@ -5,7 +5,7 @@ namespace Junges\Kafka\Contracts;
 use Junges\Kafka\Commit\Contracts\CommitterFactory;
 use Junges\Kafka\Config\Sasl;
 
-interface ConsumerBuilder
+interface ConsumerBuilder extends InteractsWithConfigCallbacks
 {
     /**
      * Creates a new ConsumerBuilder instance.
