+++ src/Producers/Producer.php
@@ -39,6 +39,10 @@ class Producer
             $conf->set($key, $value);
         }
 
+        foreach ($this->config->getConfigCallbacks() as $method => $callback) {
+            $conf->{$method}($callback);
+        }
+
         return $conf;
     }
